use module over inheritance
